+++ android/app/src/main/java/org/openmasjidos/kiosk/ui/KioskRoot.kt
@@ -92,7 +92,6 @@ fun KioskRoot(
                         noPinSet = ui.config?.pinHash?.isNotBlank() != true,
                         exitAllowed = ui.exitAllowed,
                         showPinningHint = !isDeviceOwner,
-                        onInstallAppUpdate = vm::startAppUpdate,
                         onScanReaders = vm::scanForReaders,
                         onStopReaderScan = vm::stopReaderScan,
                         onConnectReader = vm::connectReader,